:recycle:	refactor: dto명칭변경, 기본값 지정
